Make letter wrong only when it is not found
Correct the cyan rule to only declare a letter wrong placed when it is
different in the guess and correct word.

This fixes the bug that caused correctly positioned letters to be
declared wrong when there are extra instances of the same letter guessed
on the wrong positions. Now cases such as `guess=rarar, word=ardor` are
correctly evaluated to `[yellow, yellow, cyan, cyan, green]`.

Co-authored-by: Gabriel Azevedo Batalha <batalha800@gmail.com>
Co-authored-by: Davi Antônio da Silva Santos <antoniossdavi@gmail.com>